@@ -1,14 +1,14 @@
 [tool.poetry]
 name = "statute-trees"
-version = "0.0.17"
+version = "0.0.18"
 description = "Law mostly consists of tree-like nodes. This package supports a common tree format for Philippine Codifications, Statutes, and Documents, using a uniform node structure (i.e., leaves of a tree) identified by a given material path."
 authors = ["Marcelino G. Veloso III <mars@veloso.one>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.10"
-citation-utils = "^0.0.15"   # has loguru
-statute-patterns = "^0.0.23"
+python = "3.11.0"
+citation-utils = "^0.0.17"   # has loguru
+statute-patterns = "^0.0.24"
 email-validator = "^1.3.0"
 markdown = "^3.4.1"
 jinja2 = "^3.1.2"
@@ -18,23 +18,16 @@ jinja2 = "^3.1.2"
 rich = "^12.6"
 pytest = "^7.2"
 pytest-datadir = "^1.4.1"
-pytest-instafail = "^0.4.2"
 pytest-cov = "^2.12.1"
-mypy = "^0.982"
-black = "^22.10.0"
-isort = "^5.10"
-pre-commit = "^2.20"
+pre-commit = "^2.21"
 types-Markdown = "^3.4.0"
 types-PyYAML = "^6.0.7"
 types-python-slugify = "^5.0.4"
 types-python-dateutil = "^2.8.19.2"
-ipykernel = "^6.17.0"
-notebook = "^6.4.11"
-jupyter = "^1.0.0"
 
 
 [tool.pytest]
-minversion = "7.1"
+minversion = "7.2"
 addopts = "-ra -q --cov=tests/"
 testpaths = ["tests"]
 
