@@ -1,6 +1,7 @@
 import json
+from collections.abc import Iterator
 from pathlib import Path
-from typing import Iterator, Union
+from typing import Union
 
 import yaml
 from dateutil.parser import parse
@@ -20,13 +21,17 @@ from .resources import (
 
 
 class CodeUnit(Node, TreeishNode):
-    """For Codification objects. Unlike a Statute which needs to be pre-processed, a Codification is human edited. A Codification is an attempt to unify disconnected Statutes into a single entity. For instance, the `Family Code of the Philippines` is contained in Executive Order No. 209 (1987). However it has since been amended by various laws such as Republic Act No. 8533 (1998) and Republic Act No. 10572 (2013) among others. In light of the need to record a history, each Codification may contain a `history` field."""
+    """For Codification objects. Unlike a Statute which needs to be pre-processed, a Codification is human edited. A Codification is an attempt to unify disconnected Statutes into a single entity. For instance, the `Family Code of the Philippines` is contained in Executive Order No. 209 (1987). However it has since been amended by various laws such as Republic Act No. 8533 (1998) and Republic Act No. 10572 (2013) among others. In light of the need to record a history, each Codification may contain a `history` field.
+    """
 
     id: str = generic_mp
-    history: list[Union[CitationAffector, StatuteAffector]] | None = Field(
+    history: list[CitationAffector | StatuteAffector] | None = Field(
         None,
         title="Unit History",
-        description="Used in Codifications to show each statute or citation affecting the unit.",
+        description=(
+            "Used in Codifications to show each statute or citation affecting"
+            " the unit."
+        ),
     )
     units: list["CodeUnit"] | None = Field(None)
 
