@@ -2,8 +2,8 @@ import datetime
 import re
 import sys
 from abc import ABC, abstractmethod
+from collections.abc import Iterator
 from enum import Enum
-from typing import Iterator
 
 from citation_utils import extract_citations
 from dateutil.parser import parse
@@ -36,7 +36,11 @@ Note: The fields are marked with col and index for future use by the sqlpyd libr
 generic_mp = Field(
     ...,
     title="Material Path ID",
-    description="The material path of the unit within the document. All tree-like documents should start with 1.; all children of 1. will use this as a prefix, e.g. 1.1., 1.2., etc.",
+    description=(
+        "The material path of the unit within the document. All tree-like"
+        " documents should start with 1.; all children of 1. will use this as"
+        " a prefix, e.g. 1.1., 1.2., etc."
+    ),
     regex=r"^(\d+\.)+$",
     col=str,
     index=True,
@@ -44,14 +48,20 @@ generic_mp = Field(
 generic_item = Field(
     ...,
     title="Item or Locator",
-    description="Primarily determines positioning of the unit sought within a tree; e.g Section 1, Container 1, etc.",
+    description=(
+        "Primarily determines positioning of the unit sought within a tree;"
+        " e.g Section 1, Container 1, etc."
+    ),
     col=str,
     index=True,
 )
 generic_caption = Field(
     None,
     title="Caption",
-    description="When supplied, this provides additional context of the unit sought within a tree.",
+    description=(
+        "When supplied, this provides additional context of the unit sought"
+        " within a tree."
+    ),
     max_length=5000,
     col=str,
     index=True,
@@ -60,7 +70,10 @@ generic_caption = Field(
 generic_content = Field(
     None,
     title="Content",
-    description="When supplied, this provides the content proper of the unit sought within a tree.",
+    description=(
+        "When supplied, this provides the content proper of the unit sought"
+        " within a tree."
+    ),
     max_length=100000,
     col=str,
     index=True,
@@ -69,7 +82,11 @@ generic_content = Field(
 generic_variant = Field(
     None,
     title="Document / Statute / Codification Variant",
-    description="There are statutes which have the same category, id, and date, requiring a third-party key to make the value unique. This is the variant key. Example: Administrative Matter No. 19-03-24-SC.",
+    description=(
+        "There are statutes which have the same category, id, and date,"
+        " requiring a third-party key to make the value unique. This is the"
+        " variant key. Example: Administrative Matter No. 19-03-24-SC."
+    ),
     col=int,
     index=True,
 )
@@ -135,23 +152,27 @@ class Identifier(BaseModel):
 
     @property
     def slug(self):
-        """Extracts the first name of the email address, adds the year from the date and joins the details to forms a single slug."""
+        """Extracts the first name of the email address, adds the year from the date and joins the details to forms a single slug.
+        """
         authors = "-".join(email.split("@")[0] for email in self.emails)
         elements = [authors, self.date.year, self.text, f"v{self.variant}"]
-        joined_text = "-".join((str(e) for e in elements))
+        joined_text = "-".join(str(e) for e in elements)
         slug = slugify(joined_text)
         return slug
 
 
 class Page(BaseModel):
-    """HTML pages will require a `title` and a `description`. Contains the following common fields for use in tree-like structures: `created`, `modified`, `id`, `variant`, `units`."""
+    """HTML pages will require a `title` and a `description`. Contains the following common fields for use in tree-like structures: `created`, `modified`, `id`, `variant`, `units`.
+    """
 
     created: float = Field(col=float)
     modified: float = Field(col=float)
     emails: list[EmailStr] = generic_email
     id: str = Field(
         ...,
-        title="Non-traditional identifier created by slugifying certain fields.",
+        title=(
+            "Non-traditional identifier created by slugifying certain fields."
+        ),
         description="Primary key.",
         col=str,
     )
@@ -166,13 +187,16 @@ class Page(BaseModel):
     units: str = Field(
         None,
         title="Unit Tree.",
-        description="Tree in JSON string format that can be formatted via html.",
+        description=(
+            "Tree in JSON string format that can be formatted via html."
+        ),
         col=str,
     )
 
 
 class Node(BaseModel):
-    """Generic node containing the `item`, `caption` and `content` fields. Used in all tree-like structures to signify a node in the tree."""
+    """Generic node containing the `item`, `caption` and `content` fields. Used in all tree-like structures to signify a node in the tree.
+    """
 
     item: str = generic_item
     caption: str | None = generic_caption
@@ -187,7 +211,8 @@ class Node(BaseModel):
 
 
 class StatuteBase(BaseModel):
-    """Unlike a `Rule` object under `statute_patterns`, the fields for category and serial id are optional since a passed statute may not have been included in the limited set of statutes included in the `extract_rules()` function."""
+    """Unlike a `Rule` object under `statute_patterns`, the fields for category and serial id are optional since a passed statute may not have been included in the limited set of statutes included in the `extract_rules()` function.
+    """
 
     statute_category: StatuteSerialCategory | None = Field(
         None,
@@ -213,7 +238,8 @@ class StatuteBase(BaseModel):
 
     @classmethod
     def from_rule(cls, r: Rule):
-        """A rule is generated from a Path. Since there are some paths that do not map to rules because of variants, need to apply a special filter in extracting the proper `statute_serial_id`."""
+        """A rule is generated from a Path. Since there are some paths that do not map to rules because of variants, need to apply a special filter in extracting the proper `statute_serial_id`.
+        """
         target_id = r.id
         if r.cat == "rule_am" and re.search(r"^.*sc-\d+$", r.id):
             # handle cases like rule_am/00-5-03-sc-1 and rule_am/00-5-03-sc-2
@@ -222,7 +248,8 @@ class StatuteBase(BaseModel):
 
 
 class EventStatute(StatuteBase):
-    """The statute + unit item affecting a Codification unit. The combination of the `statute_category`, `statute_serial_id`, `date` and `variant` make it possible to get unique Statute documents."""
+    """The statute + unit item affecting a Codification unit. The combination of the `statute_category`, `statute_serial_id`, `date` and `variant` make it possible to get unique Statute documents.
+    """
 
     locator: str = generic_item
     caption: str | None = generic_caption
@@ -230,7 +257,11 @@ class EventStatute(StatuteBase):
     statute: str = Field(
         ...,
         title="Serial Statute Title",
-        description="Text that is parseable, see statute-patterns; will generate a `statute_category` and `statute_serial_id`. e.g. 'Republic Act No. 386' produces `ra` and `386`, respectively.",
+        description=(
+            "Text that is parseable, see statute-patterns; will generate a"
+            " `statute_category` and `statute_serial_id`. e.g. 'Republic Act"
+            " No. 386' produces `ra` and `386`, respectively."
+        ),
         min_length=4,
         col=str,
         index=True,
@@ -294,7 +325,9 @@ class StatuteAffector(EventStatute):
     action: StatuteAffectorAction = Field(
         StatuteAffectorAction.Originated,
         title="Codification Event Action",
-        description="The action of the event in relation to a Codification provision,",
+        description=(
+            "The action of the event in relation to a Codification provision,"
+        ),
         col=str,
     )
 
@@ -308,7 +341,11 @@ class FTSQuery(BaseModel):
     query: str = Field(
         ...,
         title="FTS5 SQLite-Based Expression",
-        description='e.g. "power to tax" AND ("sovereignty" OR "lifeblood" OR "security against its abuse" OR "responsibility of the legislature")',
+        description=(
+            'e.g. "power to tax" AND ("sovereignty" OR "lifeblood" OR'
+            ' "security against its abuse" OR "responsibility of the'
+            ' legislature")'
+        ),
         min_length=5,
         col=str,
     )
@@ -347,12 +384,16 @@ class EventCitation(BaseModel):
 
 
 class CitationAffector(EventCitation):
-    """Events can be sourced from a decision. The `decision_title` is represented by the `citation`. The event's effect is signified through the `action` as contextualized through supplied context in the `snippet`.."""
+    """Events can be sourced from a decision. The `decision_title` is represented by the `citation`. The event's effect is signified through the `action` as contextualized through supplied context in the `snippet`..
+    """
 
     decision_title: str = Field(
         ...,
         title="Decision Title",
-        description="Text can represent the decision as a short title, e.g. 'Tañada v. Tuvera'",
+        description=(
+            "Text can represent the decision as a short title, e.g. 'Tañada v."
+            " Tuvera'"
+        ),
         min_length=6,
         max_length=100,
         col=str,
@@ -361,7 +402,10 @@ class CitationAffector(EventCitation):
     snippet: str = Field(
         ...,
         title="Snippet of the Decision",
-        description="Text of the decision that is relevant in describing the context of the action.",
+        description=(
+            "Text of the decision that is relevant in describing the context"
+            " of the action."
+        ),
         min_length=6,
         col=str,
         index=True,
@@ -379,20 +423,27 @@ class CitationAffector(EventCitation):
 
 
 class TreeishNode(ABC):
-    """The building block of the tree. Each category of tree is different since the way they're built is nuanced, e.g. CodeUnits need a `history` field, and DocUnits need a `sources` field. However both types share the same foundational structure."""
+    """The building block of the tree. Each category of tree is different since the way they're built is nuanced, e.g. CodeUnits need a `history` field, and DocUnits need a `sources` field. However both types share the same foundational structure.
+    """
 
     @classmethod
     @abstractmethod
     def create_branches(cls, units: list[dict], parent_id: str = "1."):
-        """Each material path tree begins will eventually start with a root of `1.` so that each branch will be a material path (identified by the `id`) to the root."""
+        """Each material path tree begins will eventually start with a root of `1.` so that each branch will be a material path (identified by the `id`) to the root.
+        """
         raise NotImplementedError(
-            "Tree-based nodes must have a create_branches() function; note that each branching function for each tree category is different."
+            "Tree-based nodes must have a create_branches() function; note"
+            " that each branching function for each tree category is"
+            " different."
         )
 
     @classmethod
     @abstractmethod
     def searchables(cls, pk: str, units: list) -> Iterator[dict]:
-        """The `pk` indicated refers to the container, i.e. the foreign key Codification / Statute / Document. So every dict generated will have a unique material path with a `unit_text` that is searchable and highlightable via sqlite's FTS."""
+        """The `pk` indicated refers to the container, i.e. the foreign key Codification / Statute / Document. So every dict generated will have a unique material path with a `unit_text` that is searchable and highlightable via sqlite's FTS.
+        """
         raise NotImplementedError(
-            "Tree-based nodes must generate sqlite-compatible fts unit_text columns that is searchable and whose snippet (see sqlite's snippet() function) can be highlighted."
+            "Tree-based nodes must generate sqlite-compatible fts unit_text"
+            " columns that is searchable and whose snippet (see sqlite's"
+            " snippet() function) can be highlighted."
         )
