@@ -15,12 +15,22 @@ def id_data() -> list[dict]:
                     "units": [
                         {
                             "item": "Article 1",
-                            "content": 'This Act shall be known as the "Civil Code of the Philippines." (n)\n',
+                            "content": (
+                                'This Act shall be known as the "Civil Code of'
+                                ' the Philippines." (n)\n'
+                            ),
                             "id": "1.1.1.1",
                         },
                         {
                             "item": "Article 2",
-                            "content": "Laws shall take effect after fifteen days following the completion of their publication either in the Official Gazette or in a newspaper of general circulation in the Philippines, unless it is otherwise provided. (1a)\n",
+                            "content": (
+                                "Laws shall take effect after fifteen days"
+                                " following the completion of their"
+                                " publication either in the Official Gazette"
+                                " or in a newspaper of general circulation in"
+                                " the Philippines, unless it is otherwise"
+                                " provided. (1a)\n"
+                            ),
                             "id": "1.1.1.2",
                         },
                     ],
@@ -35,6 +45,9 @@ def id_data() -> list[dict]:
 def test_get_node_id(id_data):
     assert get_node_id(id_data, "1.1.1.1") == {
         "item": "Article 1",
-        "content": 'This Act shall be known as the "Civil Code of the Philippines." (n)\n',
+        "content": (
+            'This Act shall be known as the "Civil Code of the Philippines."'
+            " (n)\n"
+        ),
         "id": "1.1.1.1",
     }
