@@ -15,11 +15,21 @@ def raw_data() -> list[dict]:
                     "units": [
                         {
                             "item": "Article 1",
-                            "content": 'This Act shall be known as the "Civil Code of the Philippines." (n)\n',
+                            "content": (
+                                'This Act shall be known as the "Civil Code of'
+                                ' the Philippines." (n)\n'
+                            ),
                         },
                         {
                             "item": "Article 2",
-                            "content": "Laws shall take effect after fifteen days following the completion of their publication either in the Official Gazette or in a newspaper of general circulation in the Philippines, unless it is otherwise provided. (1a)\n",
+                            "content": (
+                                "Laws shall take effect after fifteen days"
+                                " following the completion of their"
+                                " publication either in the Official Gazette"
+                                " or in a newspaper of general circulation in"
+                                " the Philippines, unless it is otherwise"
+                                " provided. (1a)\n"
+                            ),
                         },
                     ],
                 }
